[refactor] RAG 파트 리팩토링
